requirements.txt
einops
requests
pytest
click
joblib
tqdm
ipykernel
notebook
ipywidgets
jupyter
flask
matplotlib
numpy-stl
mediapy
pandas
scikit-learn
black
wandb
ftfy
tensorboard
pygame
plotly
sympy
accelerate
pre-commit
diffusers
gdown
jaxtyping
typeguard
streamlit
torchinfo
frozendict